+++ src/main/java/com/been/onlinestore/controller/dto/UserRequest.java
@@ -1,14 +1,11 @@
 package com.been.onlinestore.controller.dto;
 
-import java.time.LocalDateTime;
-
 import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Pattern;
 import javax.validation.constraints.Size;
 
 import com.been.onlinestore.domain.constant.RoleType;
-import com.been.onlinestore.service.request.UserServiceRequest;
+import com.been.onlinestore.service.dto.request.UserServiceRequest;
 
 import lombok.Builder;
 
@@ -86,32 +83,6 @@ public record UserRequest() {
 		@NotBlank @Size(max = 20)
 		@Pattern(regexp = PHONE_PATTERN_REGEX, message = PHONE_PATTERN_MESSAGE)
 		String phone
-	) {
-	}
-
-	@Builder
-	public record UpdateForm(
-		@NotBlank
-		String uid,
-
-		@NotBlank
-		String name,
-
-		@NotBlank
-		String email,
-
-		@Size(max = 20)
-		String nickname,
-
-		@NotBlank @Size(max = 20)
-		@Pattern(regexp = PHONE_PATTERN_REGEX, message = PHONE_PATTERN_MESSAGE)
-		String phone,
-
-		@NotNull
-		LocalDateTime createdAt,
-
-		@NotNull
-		LocalDateTime modifiedAt
 	) {
 	}
 }
