+++ src/main/java/com/been/onlinestore/service/dto/request/CategoryServiceRequest.java
@@ -1,4 +1,4 @@
-package com.been.onlinestore.service.request;
+package com.been.onlinestore.service.dto.request;
 
 import com.been.onlinestore.domain.Category;
