+++ src/main/java/com/been/onlinestore/service/dto/response/admin/AdminCategoryResponse.java
@@ -1,6 +1,6 @@
-package com.been.onlinestore.service.response.admin;
+package com.been.onlinestore.service.dto.response.admin;
 
-import static com.been.onlinestore.service.response.JsonFormatConst.*;
+import static com.been.onlinestore.service.dto.response.JsonFormatConst.*;
 
 import java.time.LocalDateTime;
