+++ src/test/java/com/been/onlinestore/controller/OrderApiControllerTest.java
@@ -1,4 +1,4 @@
-package com.been.onlinestore.controller.api.seller;
+package com.been.onlinestore.controller.api;
 
 import static com.been.onlinestore.controller.restdocs.FieldDescription.*;
 import static com.been.onlinestore.controller.restdocs.RestDocsUtils.*;
@@ -12,6 +12,7 @@ import static org.springframework.restdocs.request.RequestDocumentation.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
 import java.util.List;
+import java.util.Map;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -28,32 +29,31 @@ import org.springframework.restdocs.payload.JsonFieldType;
 import org.springframework.security.test.context.support.WithUserDetails;
 
 import com.been.onlinestore.config.TestSecurityConfig;
+import com.been.onlinestore.controller.dto.OrderRequest;
 import com.been.onlinestore.controller.restdocs.RestDocsSupport;
 import com.been.onlinestore.controller.restdocs.RestDocsUtils;
 import com.been.onlinestore.controller.restdocs.TagDescription;
 import com.been.onlinestore.domain.constant.DeliveryStatus;
 import com.been.onlinestore.domain.constant.OrderStatus;
-import com.been.onlinestore.repository.querydsl.order.OrderSearchCondition;
 import com.been.onlinestore.service.OrderService;
-import com.been.onlinestore.service.response.OrderProductResponse;
-import com.been.onlinestore.service.response.OrderResponse;
+import com.been.onlinestore.service.dto.response.OrderProductResponse;
+import com.been.onlinestore.service.dto.response.OrderResponse;
 
 @DisplayName("API 컨트롤러 - 주문")
 @Import(TestSecurityConfig.class)
-@WebMvcTest(SellerOrderApiController.class)
-class SellerOrderApiControllerTest extends RestDocsSupport {
+@WebMvcTest(OrderApiController.class)
+class OrderApiControllerTest extends RestDocsSupport {
 
 	@MockBean
 	private OrderService orderService;
 
-	@WithUserDetails("seller")
-	@DisplayName("[API][GET] 주문 리스트 조회 + 검색 + 페이징")
+	@WithUserDetails
+	@DisplayName("[API][GET] 주문 리스트 조회 + 페이징 - 생성일 내림차순")
 	@Test
 	void test_getOrderList_withPagination() throws Exception {
 		//Given
-		long sellerId = TestSecurityConfig.SELLER_ID;
-		Long searchProductId = 1L;
-		OrderSearchCondition cond = OrderSearchCondition.of(null, searchProductId, null, null);
+		long userId = TestSecurityConfig.USER_ID;
+		long orderProductId = 1L;
 
 		String sortName = "createdAt";
 		String direction = "desc";
@@ -62,7 +62,17 @@ class SellerOrderApiControllerTest extends RestDocsSupport {
 
 		Pageable pageable = PageRequest.of(pageNumber, pageSize, Sort.by(Sort.Order.desc(sortName)));
 		OrderProductResponse orderProductResponse1 = OrderProductResponse.of(
-			4L,
+			1L,
+			"꽃무늬 셔츠",
+			12000,
+			2,
+			24000,
+			DeliveryStatus.ACCEPT,
+			3000,
+			null
+		);
+		OrderProductResponse orderProductResponse2 = OrderProductResponse.of(
+			2L,
 			"체크 셔츠",
 			23000,
 			3,
@@ -71,47 +81,24 @@ class SellerOrderApiControllerTest extends RestDocsSupport {
 			0,
 			null
 		);
-		OrderProductResponse orderProductResponse2 = OrderProductResponse.of(
+		OrderResponse response = OrderResponse.of(
 			1L,
-			"체크 셔츠",
-			23000,
-			5,
-			115000,
-			DeliveryStatus.ACCEPT,
-			0,
-			null
-		);
-		OrderResponse response1 = OrderResponse.of(
-			1L,
-			OrderResponse.OrdererResponse.of("user1", "01012345678"),
-			OrderResponse.DeliveryRequestResponse.of("서울 종로구 청와대로 1", "user1", "01012345678"),
-			List.of(orderProductResponse1),
-			85500,
+			OrderResponse.OrdererResponse.of("user", "01012345678"),
+			OrderResponse.DeliveryRequestResponse.of("서울 종로구 청와대로 1", "user", "01011112222"),
+			List.of(orderProductResponse1, orderProductResponse2),
+			93000,
 			OrderStatus.ORDER,
-			now().minusDays(1),
-			now().minusDays(1)
+			now(),
+			now()
 		);
-		OrderResponse response2 = OrderResponse.of(
-			2L,
-			OrderResponse.OrdererResponse.of("user2", "01011112222"),
-			OrderResponse.DeliveryRequestResponse.of("서울 중구 세종대로 110 서울특별시청", "user2", "01011112222"),
-			List.of(orderProductResponse2),
-			24000,
-			OrderStatus.ORDER,
-			now().minusHours(1),
-			now().minusHours(1)
-		);
-		List<OrderResponse> content = List.of(response2, response1);
+		List<OrderResponse> content = List.of(response);
 		Page<OrderResponse> page = new PageImpl<>(content, pageable, content.size());
 
-		given(orderService.findOrdersBySeller(sellerId, cond, pageable)).willReturn(page);
+		given(orderService.findOrdersByOrderer(userId, pageable)).willReturn(page);
 
 		//When & Then
-		String prefix = "검색할 ";
-
 		mvc.perform(
-				get("/api/seller/orders")
-					.queryParam("productId", String.valueOf(searchProductId))
+				get("/api/orders")
 					.queryParam("page", String.valueOf(pageNumber))
 					.queryParam("size", String.valueOf(pageSize))
 					.queryParam("sort", sortName + "," + direction)
@@ -120,31 +107,21 @@ class SellerOrderApiControllerTest extends RestDocsSupport {
 			.andExpect(content().contentTypeCompatibleWith(MediaType.APPLICATION_JSON))
 			.andExpect(jsonPath("$.status").value("success"))
 			.andExpect(jsonPath("$.data").isArray())
-			.andExpect(jsonPath("$.data[0].id").value(response2.id()))
-			.andExpect(jsonPath("$.data[0].orderer.uid").value(response2.orderer().uid()))
+			.andExpect(jsonPath("$.data[0].id").value(response.id()))
+			.andExpect(jsonPath("$.data[0].orderer.uid").value(response.orderer().uid()))
 			.andExpect(jsonPath("$.data[0].deliveryRequest.deliveryAddress").isNotEmpty())
 			.andExpect(jsonPath("$.data[0].orderProducts").isArray())
-			.andExpect(jsonPath("$.data[0].orderProducts[0].id").value(response2.orderProducts().get(0).id()))
+			.andExpect(jsonPath("$.data[0].orderProducts[0].id").value(orderProductId))
 			.andExpect(jsonPath("$.page.number").value(page.getNumber()))
 			.andExpect(jsonPath("$.page.size").value(page.getSize()))
 			.andExpect(jsonPath("$.page.totalPages").value(page.getTotalPages()))
 			.andExpect(jsonPath("$.page.totalElements").value(page.getTotalElements()))
 			.andDo(document(
-				"seller/order/getOrders-searching",
-				sellerApiDescription(TagDescription.ORDER, "주문 페이징 조회 + 검색"),
+				"user/order/getOrders",
+				userApiDescription(TagDescription.ORDER, "주문 페이징 조회"),
 				preprocessRequest(prettyPrint()),
 				preprocessResponse(prettyPrint()),
-				requestParameters(PAGE_REQUEST_PARAM)
-					.and(
-						parameterWithName("ordererId")
-							.description(prefix + ORDERER_ID.getDescription()).optional(),
-						parameterWithName("productId")
-							.description(prefix + PRODUCT_ID.getDescription()).optional(),
-						parameterWithName("deliveryStatus")
-							.description(prefix + ORDER_PRODUCT_DELIVERY_STATUS.getDescription()).optional(),
-						parameterWithName("orderStatus")
-							.description(prefix + ORDER_STATUS.getDescription()).optional()
-					),
+				requestParameters(PAGE_REQUEST_PARAM),
 				responseFields(
 					RestDocsUtils.STATUS,
 					fieldWithPath("data[].id").type(JsonFieldType.NUMBER)
@@ -185,17 +162,27 @@ class SellerOrderApiControllerTest extends RestDocsSupport {
 						.description(MODIFIED_AT.getDescription())
 				).and(PAGE_INFO)
 			));
-		then(orderService).should().findOrdersBySeller(sellerId, cond, pageable);
+		then(orderService).should().findOrdersByOrderer(userId, pageable);
 	}
 
-	@WithUserDetails("seller")
+	@WithUserDetails
 	@DisplayName("[API][GET] 주문 상세 조회")
 	@Test
 	void test_getOrder() throws Exception {
 		//Given
-		long sellerId = TestSecurityConfig.SELLER_ID;
+		long userId = TestSecurityConfig.USER_ID;
 
-		OrderProductResponse orderProductResponse = OrderProductResponse.of(
+		OrderProductResponse orderProductResponse1 = OrderProductResponse.of(
+			1L,
+			"꽃무늬 셔츠",
+			12000,
+			2,
+			24000,
+			DeliveryStatus.ACCEPT,
+			3000,
+			null
+		);
+		OrderProductResponse orderProductResponse2 = OrderProductResponse.of(
 			2L,
 			"체크 셔츠",
 			23000,
@@ -207,19 +194,18 @@ class SellerOrderApiControllerTest extends RestDocsSupport {
 		);
 		OrderResponse response = OrderResponse.of(
 			1L,
-			OrderResponse.OrdererResponse.of("user1", "01012345678"),
-			OrderResponse.DeliveryRequestResponse.of("서울 종로구 청와대로 1", "user1", "01012345678"),
-			List.of(orderProductResponse),
-			69000,
+			OrderResponse.OrdererResponse.of("user", "01012345678"),
+			OrderResponse.DeliveryRequestResponse.of("서울 종로구 청와대로 1", "user", "01011112222"),
+			List.of(orderProductResponse1, orderProductResponse2),
+			93000,
 			OrderStatus.ORDER,
-			now().minusDays(1),
-			now().minusDays(1)
+			now(),
+			now()
 		);
-
-		given(orderService.findOrderBySeller(response.id(), sellerId)).willReturn(response);
+		given(orderService.findOrderByOrderer(response.id(), userId)).willReturn(response);
 
 		//When & Then
-		mvc.perform(get("/api/seller/orders/{orderId}", response.id()))
+		mvc.perform(get("/api/orders/{orderId}", response.id()))
 			.andExpect(status().isOk())
 			.andExpect(content().contentTypeCompatibleWith(MediaType.APPLICATION_JSON))
 			.andExpect(jsonPath("$.status").value("success"))
@@ -229,12 +215,12 @@ class SellerOrderApiControllerTest extends RestDocsSupport {
 			.andExpect(jsonPath("$.data.orderProducts").isArray())
 			.andExpect(jsonPath("$.data.orderProducts[0].id").value(response.orderProducts().get(0).id()))
 			.andDo(document(
-				"seller/order/getOrder",
-				sellerApiDescription(TagDescription.ORDER, "주문 상세 조회"),
+				"user/order/getOrder",
+				userApiDescription(TagDescription.ORDER, "주문 상세 조회"),
 				preprocessRequest(prettyPrint()),
 				preprocessResponse(prettyPrint()),
 				pathParameters(
-					parameterWithName("orderId").description(PRODUCT_ID.getDescription())
+					parameterWithName("orderId").description(ORDER_ID.getDescription())
 				),
 				responseFields(
 					RestDocsUtils.STATUS,
@@ -276,5 +262,86 @@ class SellerOrderApiControllerTest extends RestDocsSupport {
 						.description(MODIFIED_AT.getDescription())
 				)
 			));
+		then(orderService).should().findOrderByOrderer(response.id(), userId);
+	}
+
+	@WithUserDetails
+	@DisplayName("[API][POST] 주문하기")
+	@Test
+	void test_order() throws Exception {
+		//Given
+		long orderId = 1L;
+		long userId = TestSecurityConfig.USER_ID;
+		OrderRequest request = new OrderRequest(Map.of(1L, 10), "서울 종로구 청와대로 1", "user", "01011112222");
+
+		given(orderService.order(userId, request.toServiceRequest())).willReturn(orderId);
+
+		//When & Then
+		mvc.perform(
+				post("/api/orders")
+					.contentType(MediaType.APPLICATION_JSON)
+					.accept(MediaType.APPLICATION_JSON)
+					.characterEncoding("UTF-8")
+					.content(mapper.writeValueAsString(request))
+			)
+			.andExpect(status().isOk())
+			.andExpect(content().contentTypeCompatibleWith(MediaType.APPLICATION_JSON))
+			.andExpect(jsonPath("$.status").value("success"))
+			.andExpect(jsonPath("$.data.id").value(orderId))
+			.andDo(document(
+				"user/order/order",
+				userApiDescription(TagDescription.ORDER, "주문하기"),
+				preprocessRequest(prettyPrint()),
+				preprocessResponse(prettyPrint()),
+				requestFields(
+					subsectionWithPath("productIdToQuantity").type(JsonFieldType.OBJECT)
+						.description("상품 시퀀스 : 수량 맵"),
+					fieldWithPath("deliveryAddress").type(JsonFieldType.STRING)
+						.description(DELIVERY_REQUEST_ADDRESS.getDescription()),
+					fieldWithPath("receiverName").type(JsonFieldType.STRING)
+						.description(DELIVERY_REQUEST_RECEIVER_NAME.getDescription()),
+					fieldWithPath("receiverPhone").type(JsonFieldType.STRING)
+						.description(DELIVERY_REQUEST_RECEIVER_PHONE.getDescription())
+				),
+				responseFields(
+					RestDocsUtils.STATUS,
+					fieldWithPath("data.id").type(JsonFieldType.NUMBER)
+						.description(ADD.getDescription() + ORDER_ID.getDescription())
+				)
+			));
+		then(orderService).should().order(userId, request.toServiceRequest());
+	}
+
+	@WithUserDetails
+	@DisplayName("[API][PUT] 주문 취소")
+	@Test
+	void test_cancelOrder() throws Exception {
+		//Given
+		long orderId = 1L;
+		long userId = TestSecurityConfig.USER_ID;
+
+		given(orderService.cancelOrder(orderId, userId)).willReturn(orderId);
+
+		//When & Then
+		mvc.perform(put("/api/orders/{orderId}/cancel", orderId))
+			.andExpect(status().isOk())
+			.andExpect(content().contentTypeCompatibleWith(MediaType.APPLICATION_JSON))
+			.andExpect(jsonPath("$.status").value("success"))
+			.andExpect(jsonPath("$.data.id").value(orderId))
+			.andDo(document(
+				"user/order/cancelOrder",
+				userApiDescription(TagDescription.ORDER, "주문 취소하기"),
+				preprocessRequest(prettyPrint()),
+				preprocessResponse(prettyPrint()),
+				pathParameters(
+					parameterWithName("orderId").description(ORDER_ID.getDescription())
+				),
+				responseFields(
+					RestDocsUtils.STATUS,
+					fieldWithPath("data.id").type(JsonFieldType.NUMBER)
+						.description("취소된 " + ORDER_ID.getDescription())
+				)
+			));
+		then(orderService).should().cancelOrder(orderId, userId);
 	}
 }
