+++ src/test/java/com/been/onlinestore/service/OrderServiceTest.java
@@ -30,8 +30,8 @@ import com.been.onlinestore.repository.OrderRepository;
 import com.been.onlinestore.repository.ProductRepository;
 import com.been.onlinestore.repository.UserRepository;
 import com.been.onlinestore.repository.querydsl.order.OrderSearchCondition;
-import com.been.onlinestore.service.request.OrderServiceRequest;
-import com.been.onlinestore.service.response.OrderResponse;
+import com.been.onlinestore.service.dto.request.OrderServiceRequest;
+import com.been.onlinestore.service.dto.response.OrderResponse;
 
 @DisplayName("비즈니스 로직 - 주문")
 @ExtendWith(MockitoExtension.class)
