+++ src/test/java/com/been/onlinestore/util/ProductTestDataUtil.java
@@ -9,7 +9,7 @@ import org.springframework.test.util.ReflectionTestUtils;
 import com.been.onlinestore.domain.Product;
 import com.been.onlinestore.domain.constant.SaleStatus;
 import com.been.onlinestore.repository.querydsl.product.AdminProductResponse;
-import com.been.onlinestore.service.response.CategoryProductResponse;
+import com.been.onlinestore.service.dto.response.CategoryProductResponse;
 
 public class ProductTestDataUtil {
